+++ enro-core/src/main/java/dev/enro/core/result/internal/PendingResult.kt
@@ -1,19 +1,23 @@
 package dev.enro.core.result.internal
 
+import dev.enro.core.AnyOpenInstruction
 import dev.enro.core.NavigationKey
 import kotlin.reflect.KClass
 
 internal sealed class PendingResult {
     abstract val resultChannelId: ResultChannelId
+    abstract val instruction: AnyOpenInstruction
     abstract val navigationKey: NavigationKey.WithResult<*>
 
     class Closed(
         override val resultChannelId: ResultChannelId,
+        override val instruction: AnyOpenInstruction,
         override val navigationKey: NavigationKey.WithResult<*>,
-    ): PendingResult()
+    ) : PendingResult()
 
     data class Result(
         override val resultChannelId: ResultChannelId,
+        override val instruction: AnyOpenInstruction,
         override val navigationKey: NavigationKey.WithResult<*>,
         val resultType: KClass<out Any>,
         val result: Any